requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7

# AWS Integration
boto3==1.34.10
botocore==1.34.10

# Environment Configuration
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0

# Utilities
python-dateutil==2.8.2
